+++ adetailer/args.py
@@ -56,6 +56,8 @@ class ArgsList(UserList):
 class ADetailerArgs(BaseModel, extra=Extra.forbid):
     ad_model: str = "None"
     ad_model_classes: str = ""
+    ad_model_classes_exclude: bool = False
+    ad_model_classes_excluded: str = ""
     ad_tab_enable: bool = True
     ad_prompt: str = ""
     ad_negative_prompt: str = ""
@@ -129,6 +131,8 @@ class ADetailerArgs(BaseModel, extra=Extra.forbid):
         ppop = partial(self.ppop, p)
 
         ppop("ADetailer model classes")
+        ppop("ADetailer classes exclude")
+        ppop("ADetailer model classes excluded")
         ppop("ADetailer prompt")
         ppop("ADetailer negative prompt")
         p.pop("ADetailer tab enable", None)  # always pop
@@ -218,6 +222,8 @@ class ADetailerArgs(BaseModel, extra=Extra.forbid):
 _all_args = [
     ("ad_model", "ADetailer model"),
     ("ad_model_classes", "ADetailer model classes"),
+    ("ad_model_classes_exclude", "ADetailer classes exclude"),
+    ("ad_model_classes_excluded", "ADetailer model classes excluded"),
     ("ad_tab_enable", "ADetailer tab enable"),
     ("ad_prompt", "ADetailer prompt"),
     ("ad_negative_prompt", "ADetailer negative prompt"),